trading-app/pyproject.toml
[project]
name = "trading-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.116.1",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.6.1",
  "pyzmq>=27.0.1",
  "streamlit>=1.48.0",
  "streamlit-shadcn-ui>=0.1.18",
  "uvicorn>=0.35.0",
]

[dependency-groups]
dev = ["black>=25.1.0", "pre-commit>=4.2.0", "pytest>=8.4.1", "ruff>=0.12.8"]

[tool.uv]
package = true

[project.scripts]
run-trading-app = "trading_app.entrypoint:run"
